+++ utoipa-gen/src/path.rs
@@ -1,6 +1,6 @@
 use std::{io::Error, str::FromStr};
 
-use proc_macro2::{Group, Ident, Span, TokenStream as TokenStream2};
+use proc_macro2::{Group, Ident, TokenStream as TokenStream2};
 use proc_macro_error::{abort_call_site, OptionExt, ResultExt};
 use quote::{format_ident, quote, ToTokens};
 use syn::{
@@ -8,16 +8,22 @@ use syn::{
     parse::{Parse, ParseStream},
     parse2,
     punctuated::Punctuated,
-    token::Bracket,
-    LitBool, LitStr, Token,
+    LitStr, Token,
 };
 
-use crate::{
-    component_type::{ComponentFormat, ComponentType},
+use crate::ext::{Argument, ArgumentIn};
+use crate::Deprecated;
+
+use self::{
+    parameter::{Parameter, ParameterIn},
     request_body::RequestBodyAttr,
     response::{Response, Responses},
-    Deprecated, Required,
 };
+
+pub mod parameter;
+mod property;
+mod request_body;
+mod response;
 
 const PATH_STRUCT_PREFIX: &str = "__path_";
 
@@ -60,7 +66,61 @@ pub struct PathAttr {
     path: Option<String>,
     operation_id: Option<String>,
     tag: Option<String>,
-    pub params: Option<Vec<Parameter>>,
+    params: Option<Vec<Parameter>>,
+}
+
+impl PathAttr {
+    #[cfg(feature = "actix_extras")]
+    pub fn update_parameters(&mut self, arguments: Option<Vec<Argument>>) {
+        if let Some(arguments) = arguments {
+            let new_parameter = |argument: &Argument| {
+                Parameter::new(
+                    &argument.name.as_ref().unwrap(),
+                    argument.ident,
+                    if argument.argument_in == ArgumentIn::Path {
+                        ParameterIn::Path
+                    } else {
+                        ParameterIn::Query
+                    },
+                )
+            };
+
+            if let Some(ref mut parameters) = self.params {
+                parameters.iter_mut().for_each(|parameter| {
+                    if let Some(argument) = arguments
+                        .iter()
+                        .find(|argument| argument.name.as_ref() == Some(&parameter.name))
+                    {
+                        parameter.update_parameter_type(argument.ident)
+                    }
+                });
+
+                // add argument to the parameters if argument has a name and it does not exists in parameters
+                arguments
+                    .iter()
+                    .filter(|argument| argument.has_name())
+                    .for_each(|argument| {
+                        // cannot use filter() for mutli borrow situation. :(
+                        if !parameters
+                            .iter()
+                            .any(|parameter| Some(&parameter.name) == argument.name.as_ref())
+                        {
+                            // if parameters does not contain argument
+                            parameters.push(new_parameter(argument))
+                        }
+                    });
+            } else {
+                // no parameters at all, add arguments to the parameters if argument has a name
+                let mut params = Vec::with_capacity(arguments.len());
+                arguments
+                    .iter()
+                    .filter(|argument| argument.has_name())
+                    .map(new_parameter)
+                    .for_each(|parameter| params.push(parameter));
+                self.params = Some(params);
+            }
+        }
+    }
 }
 
 impl Parse for PathAttr {
@@ -240,210 +300,6 @@ impl ToTokens for PathOperation {
         tokens.extend(path_item_type);
     }
 }
-
-/// Parameter of request suchs as in path, header, query or cookie
-///
-/// For example path `/users/{id}` the path parameter is used to define
-/// type, format and other details of the `{id}` parameter within the path
-///
-/// Parse is executed for following formats:
-///
-/// * ("id" = String, path, required, deprecated, description = "Users database id"),
-/// * ("id", path, required, deprecated, description = "Users database id"),
-///
-/// The `= String` type statement is optional if automatic resolvation is supported.
-#[derive(Default)]
-#[cfg_attr(feature = "debug", derive(Debug))]
-pub struct Parameter {
-    pub name: String,
-    parameter_in: ParameterIn,
-    required: bool,
-    deprecated: bool,
-    pub parameter_type: Option<Ident>,
-    is_array: bool,
-    description: Option<String>,
-}
-
-impl Parameter {
-    pub fn new<S: AsRef<str>>(name: S, parameter_type: &Ident, parameter_in: ParameterIn) -> Self {
-        let required = parameter_in == ParameterIn::Path;
-
-        Self {
-            name: name.as_ref().to_string(),
-            parameter_type: Some(parameter_type.clone()),
-            parameter_in,
-            required,
-            ..Default::default()
-        }
-    }
-
-    pub fn update_parameter_type(&mut self, ident: &Ident) {
-        self.parameter_type = Some(ident.clone());
-    }
-}
-
-impl Parse for Parameter {
-    fn parse(input: ParseStream) -> syn::Result<Self> {
-        let mut parameter = Parameter::default();
-
-        if input.peek(LitStr) {
-            // parse name
-            let name = input.parse::<LitStr>().unwrap().value();
-            parameter.name = name;
-            if input.peek(Token![=]) && input.peek2(syn::Ident) {
-                // parse type for name if provided
-                input.parse::<Token![=]>().unwrap();
-                parameter.parameter_type = Some(input.parse::<syn::Ident>().unwrap());
-            } else if input.peek(Token![=]) && input.peek2(Bracket) {
-                // parse group as array
-                input.parse::<Token![=]>().unwrap();
-                parameter.is_array = true;
-                let group_content;
-                bracketed!(group_content in input);
-                parameter.parameter_type = Some(group_content.parse::<Ident>().unwrap());
-            }
-        } else {
-            return Err(input.error("expected first element to be LitStr parameter name"));
-        }
-        if input.peek(Token![,]) {
-            input.parse::<Token![,]>().unwrap();
-        }
-
-        let parse_bool = |input: ParseStream| {
-            // support assign form as: required = bool
-            if input.peek(Token![=]) && input.peek2(LitBool) {
-                input.parse::<Token![=]>().unwrap();
-
-                input.parse::<LitBool>().unwrap().value()
-            } else {
-                // quick form as: required
-                true
-            }
-        };
-
-        loop {
-            let ident = input.parse::<syn::Ident>().unwrap();
-            let name = &*ident.to_string();
-            match name {
-                "path" | "query" | "header" | "cookie" => {
-                    parameter.parameter_in = name.parse::<ParameterIn>().unwrap_or_abort();
-                    if parameter.parameter_in == ParameterIn::Path {
-                        parameter.required = true; // all path parameters are required by default
-                    }
-                }
-                "required" => parameter.required = parse_bool(input),
-                "deprecated" => parameter.deprecated = parse_bool(input),
-                "description" => {
-                    if input.peek(Token![=]) {
-                        input.parse::<Token![=]>().unwrap();
-                    }
-
-                    parameter.description = Some(
-                        input
-                            .parse::<LitStr>()
-                            .expect_or_abort("expected description value as LitStr")
-                            .value(),
-                    )
-                }
-                _ => return Err(input.error(&format!("unexpected element: {}", name))),
-            }
-
-            if input.peek(Token![,]) {
-                input.parse::<Token![,]>().unwrap();
-            }
-            if input.is_empty() {
-                break;
-            }
-        }
-        Ok(parameter)
-    }
-}
-impl ToTokens for Parameter {
-    fn to_tokens(&self, tokens: &mut TokenStream2) {
-        let name = &*self.name;
-        tokens.extend(quote! { utoipa::openapi::path::Parameter::new(#name) });
-        let parameter_in = &self.parameter_in;
-        tokens.extend(quote! { .with_in(#parameter_in) });
-
-        let required: Required = self.required.into();
-        tokens.extend(quote! { .with_required(#required) });
-
-        let deprecated: Deprecated = self.deprecated.into();
-        tokens.extend(quote! { .with_deprecated(#deprecated) });
-
-        if let Some(ref description) = self.description {
-            tokens.extend(quote! { .with_description(#description) });
-        }
-
-        if let Some(ref parameter_type) = self.parameter_type {
-            // TODO unify this property logic with the one in component.rs
-            let component_type = ComponentType(parameter_type);
-            let mut property = quote! {
-                utoipa::openapi::Property::new(
-                    #component_type
-                )
-            };
-            let format = ComponentFormat(parameter_type);
-            if format.is_known_format() {
-                property.extend(quote! {
-                    .with_format(#format)
-                })
-            }
-            if self.is_array {
-                property.extend(quote! { .to_array() });
-            }
-
-            tokens.extend(quote! { .with_schema(#property) });
-        }
-    }
-}
-
-#[cfg_attr(feature = "debug", derive(Debug))]
-#[derive(PartialEq)]
-pub enum ParameterIn {
-    Query,
-    Path,
-    Header,
-    Cookie,
-}
-
-impl Default for ParameterIn {
-    fn default() -> Self {
-        Self::Path
-    }
-}
-
-impl FromStr for ParameterIn {
-    type Err = syn::Error;
-
-    fn from_str(s: &str) -> Result<Self, Self::Err> {
-        match s {
-            "path" => Ok(Self::Path),
-            "query" => Ok(Self::Query),
-            "header" => Ok(Self::Header),
-            "cookie" => Ok(Self::Cookie),
-            _ => Err(syn::Error::new(
-                Span::call_site(),
-                &format!(
-                    "unexpected str: {}, expected one of: path, query, header, cookie",
-                    s
-                ),
-            )),
-        }
-    }
-}
-
-impl ToTokens for ParameterIn {
-    fn to_tokens(&self, tokens: &mut TokenStream2) {
-        tokens.extend(match self {
-            Self::Path => quote! { utoipa::openapi::path::ParameterIn::Path },
-            Self::Query => quote! { utoipa::openapi::path::ParameterIn::Query },
-            Self::Header => quote! { utoipa::openapi::path::ParameterIn::Header },
-            Self::Cookie => quote! { utoipa::openapi::path::ParameterIn::Cookie },
-        })
-    }
-}
-
 pub struct Path {
     path_attr: PathAttr,
     fn_name: String,
